importsdata.py
import logging, time, os, asyncio
from random import randrange
from typing import ContextManager

from telegram import Update
from telegram.ext import Updater, CommandHandler, CallbackContext, MessageHandler, Filters
from telegram.files.inputmedia import InputMediaPhoto

logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO
)

logger = logging.getLogger(__name__)